pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "cricket-agent"
version = "0.1.0"
description = "Natural language interface for querying cricket statistics"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary",
    "pyyaml",
    "requests",
    "python-dotenv",
    "anthropic",
    "typer",
    "rich",
    "fastapi>=0.129.2",
    "uvicorn[standard]>=0.41.0",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "hvac>=2.0.0",
    "spiffe>=0.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
